Proof-of-concept escape scanner using subtract-with-borrow
This scanner was derived using superoptimization, so I do not have
a good intuitive explanation for each individual operation. However,
it computes the same function as the original escape scanner, and
cuts the latency of the critical path roughly in half.

This results in a 5% speedup on the minification benchmarks when
building with clang, and no improvement on other benchmarks, or
when using gcc.